restructured the src folder
